# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sistema IoT Riego Agrícola
  version: 0.1.0
paths:
...
  /api/v1/clients/me/notification-settings:
    get:
      tags:
      - Clients
      summary: Get My Notification Settings
      operationId: get_my_notification_settings_api_v1_clients_me_notification_settings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientNotificationSettingsResponse'
      security:
      - OAuth2PasswordBearer: []
    patch:
      tags:
      - Clients
      summary: Update My Notification Settings
      operationId: update_my_notification_settings_api_v1_clients_me_notification_settings_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientNotificationSettingsUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientNotificationSettingsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/clients:
    get:
      tags:
      - Clients
      summary: List Clients
      operationId: list_clients_api_v1_clients_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Per Page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_ClientResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Clients
      summary: Create Client
      operationId: create_client_api_v1_clients_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/clients/{client_id}:
    get:
      tags:
      - Clients
      summary: Get Client
      operationId: get_client_api_v1_clients__client_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: integer
          title: Client Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Clients
      summary: Update Client
      operationId: update_client_api_v1_clients__client_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: integer
          title: Client Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Clients
      summary: Delete Client
      operationId: delete_client_api_v1_clients__client_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: integer
          title: Client Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/notification-preferences:
    get:
      tags:
      - Notification Preferences
      summary: List Notification Preferences
      operationId: list_notification_preferences_api_v1_notification_preferences_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 50
          title: Per Page
      - name: client_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Client Id
      - name: irrigation_area_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Irrigation Area Id
      - name: alert_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Alert Type
      - name: severity
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
      - name: channel
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Channel
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_NotificationPreferenceResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notification-preferences/bulk:
    put:
      tags:
      - Notification Preferences
      summary: Bulk Upsert Notification Preferences
      operationId: bulk_upsert_notification_preferences_api_v1_notification_preferences_bulk_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationPreferenceBulkUpsertRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferenceBulkUpsertResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    ClientCreate:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
        full_name:
          type: string
          title: Full Name
        company_name:
          type: string
          title: Company Name
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
      type: object
      required:
      - email
      - password
      - full_name
      - company_name
      title: ClientCreate
    ClientNotificationSettingsResponse:
      properties:
        notifications_enabled:
          type: boolean
          title: Notifications Enabled
      type: object
      required:
      - notifications_enabled
      title: ClientNotificationSettingsResponse
    ClientNotificationSettingsUpdate:
      properties:
        notifications_enabled:
          type: boolean
          title: Notifications Enabled
      type: object
      required:
      - notifications_enabled
      title: ClientNotificationSettingsUpdate
    ClientResponse:
      properties:
        id:
          type: integer
          title: Id
        user_id:
          type: integer
          title: User Id
        company_name:
          type: string
          title: Company Name
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        notifications_enabled:
          type: boolean
          title: Notifications Enabled
        user:
          anyOf:
          - $ref: '#/components/schemas/UserResponse'
          - type: 'null'
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - company_name
      - notifications_enabled
      - created_at
      - updated_at
      title: ClientResponse
    ClientUpdate:
      properties:
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        company_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Company Name
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
      type: object
      title: ClientUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NotificationPreferenceBulkUpsertRequest:
      properties:
        items:
          items:
            $ref: '#/components/schemas/NotificationPreferenceUpsertItem'
          type: array
          title: Items
      type: object
      required:
      - items
      title: NotificationPreferenceBulkUpsertRequest
    NotificationPreferenceBulkUpsertResponse:
      properties:
        created:
          type: integer
          title: Created
        updated:
          type: integer
          title: Updated
        data:
          items:
            $ref: '#/components/schemas/NotificationPreferenceResponse'
          type: array
          title: Data
      type: object
      required:
      - created
      - updated
      - data
      title: NotificationPreferenceBulkUpsertResponse
    NotificationPreferenceResponse:
      properties:
        id:
          type: integer
          title: Id
        client_id:
          type: integer
          title: Client Id
        irrigation_area_id:
          type: integer
          title: Irrigation Area Id
        alert_type:
          type: string
          title: Alert Type
        severity:
          type: string
          title: Severity
        channel:
          type: string
          title: Channel
        enabled:
          type: boolean
          title: Enabled
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - client_id
      - irrigation_area_id
      - alert_type
      - severity
      - channel
      - enabled
      - created_at
      - updated_at
      title: NotificationPreferenceResponse
    NotificationPreferenceUpsertItem:
      properties:
        irrigation_area_id:
          type: integer
          title: Irrigation Area Id
        alert_type:
          type: string
          title: Alert Type
        severity:
          type: string
          title: Severity
        channel:
          type: string
          title: Channel
        enabled:
          type: boolean
          title: Enabled
      type: object
      required:
      - irrigation_area_id
      - alert_type
      - severity
      - channel
      - enabled
      title: NotificationPreferenceUpsertItem
...
    PaginatedResponse_ClientResponse_:
      properties:
        page:
          type: integer
          title: Page
        per_page:
          type: integer
          title: Per Page
        total:
          type: integer
          title: Total
        data:
          items:
            $ref: '#/components/schemas/ClientResponse'
          type: array
          title: Data
      type: object
      required:
      - page
      - per_page
      - total
      - data
      title: PaginatedResponse[ClientResponse]
...
    PaginatedResponse_NotificationPreferenceResponse_:
      properties:
        page:
          type: integer
          title: Page
        per_page:
          type: integer
          title: Per Page
        total:
          type: integer
          title: Total
        data:
          items:
            $ref: '#/components/schemas/NotificationPreferenceResponse'
          type: array
          title: Data
      type: object
      required:
      - page
      - per_page
      - total
      - data
      title: PaginatedResponse[NotificationPreferenceResponse]
...
    UserResponse:
      properties:
        id:
          type: integer
          title: Id
        email:
          type: string
          title: Email
        full_name:
          type: string
          title: Full Name
        role:
          type: string
          title: Role
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - email
      - full_name
      - role
      - is_active
      - created_at
      - updated_at
      title: UserResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login